Made the data preprocessing function into its own module.
